requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astropy==6.0.1
astropy-iers-data==0.2026.3.9.0.47.45
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
dgl==2.4.0+cu118
distro==1.9.0
einops==0.7.0
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
ftfy==6.3.1
GPUtil==1.4.0
h11==0.16.0
healpy==1.17.3
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
objaverse==0.1.7
omegaconf==2.3.0
open-clip-torch==2.24.0
openai==1.60.0
packaging==26.0
pandas==2.2.2
par==1.3.4
param==2.2.1
parm==1.11
pillow==11.3.0
protobuf==6.33.5
psutil==7.2.2
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyerfa==2.0.1.5
pyglet==1.5.28
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scipy==1.13.1
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
timm==1.0.25
torch==2.2.2+cu118
torch.redstone==0.0.6
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.67.3
trimesh==4.3.1
triton==2.2.0
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
zipp==3.23.0
